pyproject.toml
[project]
name = "fastapi-auth-2-0"
version = "0.1.0"
description = "authentication app with some improve"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic==1.13.2",
    "annotated-types==0.7.0",
    "anyio==4.4.0",
    "certifi==2024.8.30",
    "cffi==1.17.1",
    "click==8.1.7",
    "colorama==0.4.6",
    "cryptography==43.0.3",
    "dnspython==2.7.0",
    "ecdsa==0.19.0",
    "email-validator==2.2.0",
    "fastapi-cli==0.0.5",
    "fastapi==0.112.2",
    "greenlet==3.0.3",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httptools==0.6.4",
    "httpx==0.27.2",
    "idna==3.7",
    "jinja2==3.1.4",
    "mako==1.3.5",
    "markdown-it-py==3.0.0",
    "markupsafe==2.1.5",
    "mdurl==0.1.2",
    "passlib==1.7.4",
    "psycopg2-binary==2.9.9",
    "psycopg==3.2.3",
    "pyasn1==0.6.0",
    "pycparser==2.22",
    "pydantic-core==2.20.1",
    "pydantic-settings==2.4.0",
    "pydantic==2.8.2",
    "pygments==2.18.0",
    "pyjwt==2.9.0",
    "python-dotenv==1.0.1",
    "python-jose==3.3.0",
    "python-multipart==0.0.17",
    "pyyaml==6.0.2",
    "rich==13.9.4",
    "rsa==4.9",
    "scrypt==0.8.27",
    "shellingham==1.5.4",
    "six==1.16.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.32",
    "sqlmodel==0.0.21",
    "starlette==0.38.2",
    "typer==0.13.1",
    "typing-extensions==4.12.2",
    "tzdata==2024.2",
    "uvicorn==0.30.6",
    "watchfiles==0.24.0",
    "websockets==14.1",
]
